src/height_estimation.py
"""
Height Estimation for "Harmonies" Board Game
"""

import cv2
import numpy as np

def depth_estimation(image):
    """
    Placeholder function for depth estimation methods.
    
    Parameters:
    image (numpy.ndarray): Input image in BGR format.
    
    Returns:
    numpy.ndarray: Depth map of the image.
    """
    # Placeholder implementation
    depth_map = np.zeros_like(image[:, :, 0])
    
    return depth_map

def detect_overlaps(contours):
    """
    Detects overlaps in contours to infer stacking height.
    
    Parameters:
    contours (list): List of detected contours.
    
    Returns:
    list: List of contours with inferred stacking height.
    """
    heights = []
    
    for contour in contours:
        # Placeholder logic for detecting overlaps
        height = 1  # Default height
        heights.append((contour, height))
    
    return heights

def vertical_alignment_check(contours):
    """
    Checks vertical alignment of contours to count stacked pieces.
    
    Parameters:
    contours (list): List of detected contours.
    
    Returns:
    list: List of contours with counted stacked pieces.
    """
    stacked_pieces = []
    
    for contour in contours:
        # Placeholder logic for vertical alignment check
        stacked_count = 1  # Default count
        stacked_pieces.append((contour, stacked_count))
    
    return stacked_pieces
